improve keepalive, make test infra public
