+++ src/tests/example.rs
@@ -9,11 +9,13 @@ use bevy_app::App;
 use glam::DVec3;
 use valence_client::movement::PositionAndOnGroundC2s;
 use valence_client::Client;
+use valence_core::Server;
 use valence_entity::Position;
 use valence_inventory::packet::{InventoryS2c, OpenScreenS2c};
 use valence_inventory::{Inventory, InventoryKind, OpenInventory};
 
-use super::*;
+use crate::DefaultPlugins;
+use crate::testing::scenario_single_client;
 
 /// The server's tick should increment every update.
 #[test]
@@ -63,7 +65,7 @@ fn example_test_open_inventory() {
 
     // Process a tick to get past the "on join" logic.
     app.update();
-    client_helper.clear_sent();
+    client_helper.clear_received();
 
     // Open the inventory.
     let open_inventory = OpenInventory::new(inventory_ent);
